FIX: reset listener when stopping
Although we try to reset the Listener object back to the beginnning when
stoping it, the device entities will have lost their "implicitListener"
on the "newListener" event (because reasons...)

In effect, if you want to relisten to events, you are far better
starting with a fresh Sonos instance than reusing one. But at least this
will stop the pollution of old instances on the listener chain.